project-assignment3/requirements.txt
attrs==21.4.0
avro-python3==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
configparser==5.2.0
confluent-kafka==1.8.2
fastavro==1.4.10
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
jsonschema==4.4.0
numpy==1.21.6
pandas==1.3.5
protobuf==3.20.0
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1    
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0